requirements.txt
alembic==1.16.4
beautifulsoup4==4.14.2
black==24.4.2
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.1
folium>=0.16
geopy>=2.4
google-cloud-storage==2.17.0
gunicorn==22.0.0
hypercorn>=0.15.0
h2>=4.0.0
mysql-connector-python==8.1.0
openpyxl==3.1.5
pandas==2.3.1
PyMySQL==1.1.1
psycopg2-binary==2.9.9
pytest==8.2.1
python-dotenv==1.1.1
requests==2.32.4
Authlib==1.3.1
Werkzeug==3.1.3
email-validator==2.1.1

# SQLAlchemy/model deps assumed to be in your existing db module if needed

redis>=5

Flask-Session==0.5.0
celery==5.3.1
